children-playground-project/requirements.txt
absl-py==1.4.0
asgiref==3.7.2
astunparse==1.6.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
Django==4.2.2
django-cors-headers==4.1.0
EasyProcess==1.1
entrypoint2==1.1
flatbuffers==23.5.26
fonttools==4.40.0
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
idna==3.4
keras==2.13.1rc0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
mediapipe==0.10.1
MouseInfo==0.1.3
mss==9.0.1
numpy==1.25.0
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Quartz==9.2
pyparsing==3.1.0
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
pytweening==1.0.7
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
rubicon-objc==0.4.6
six==1.16.0
sounddevice==0.4.6
sqlparse==0.4.4
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0rc1
tensorflow-estimator==2.13.0rc0
tensorflow-macos==2.13.0rc1
termcolor==2.3.0
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
